+++ packages/orq-rc/src/orq_ai_sdk/deployments.py
@@ -1137,6 +1137,7 @@ class Deployments(BaseSDK):
                 ),
                 sentinel="[DONE]",
                 client_ref=self,
+                data_required=False,
             )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = utils.stream_to_text(http_res)
@@ -1325,6 +1326,7 @@ class Deployments(BaseSDK):
                 ),
                 sentinel="[DONE]",
                 client_ref=self,
+                data_required=False,
             )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)