+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKEEKG1SB26HECWTN019YBZX"
+        "01KKGBFEPGWE9VP1XZZHJK5JQP"
     )
     r"""The unique identifier of the data source"""