+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -8,9 +8,15 @@ from orq_ai_sdk.types import (
     UNSET,
     UNSET_SENTINEL,
 )
-from orq_ai_sdk.utils import FieldMetadata, QueryParamMetadata, get_discriminator
+from orq_ai_sdk.utils import (
+    FieldMetadata,
+    QueryParamMetadata,
+    get_discriminator,
+    validate_const,
+)
 import pydantic
 from pydantic import Discriminator, Tag, model_serializer
+from pydantic.functional_validators import AfterValidator
 from typing import Any, Dict, List, Literal, Optional, Union
 from typing_extensions import Annotated, NotRequired, TypeAliasType, TypedDict
 
@@ -79,7 +85,7 @@ class DeploymentsParametersTypedDict(TypedDict):
     type: DeploymentsDeploymentsResponseType
     properties: Dict[str, Any]
     required: NotRequired[List[str]]
-    additional_properties: NotRequired[bool]
+    additional_properties: Literal[False]
 
 
 class DeploymentsParameters(BaseModel):
@@ -94,9 +100,10 @@ class DeploymentsParameters(BaseModel):
 
     required: Optional[List[str]] = None
 
-    additional_properties: Annotated[
-        Optional[bool], pydantic.Field(alias="additionalProperties")
-    ] = None
+    ADDITIONAL_PROPERTIES: Annotated[
+        Annotated[Optional[Literal[False]], AfterValidator(validate_const(False))],
+        pydantic.Field(alias="additionalProperties"),
+    ] = False
 
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):