+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKEEKG1P7QE4J97N2G7G9MB6"
+        "01KKGBFEPEERX1B42SFA1YH875"
     )
     r"""The unique identifier of the data source"""