+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKEEKG1QG3FECPGW9H6AM3VH"
+        "01KKGBFEPG0TP111CTC0C1YQT5"
     )
     r"""The unique identifier of the data source"""