+++ packages/orq-rc/src/orq_ai_sdk/responses.py
@@ -199,6 +199,7 @@ class Responses(BaseSDK):
                 ),
                 sentinel="[DONE]",
                 client_ref=self,
+                data_required=False,
             )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = utils.stream_to_text(http_res)
@@ -390,6 +391,7 @@ class Responses(BaseSDK):
                 ),
                 sentinel="[DONE]",
                 client_ref=self,
+                data_required=False,
             )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)