+++ packages/orq-rc/RELEASES.md
@@ -6489,3 +6489,13 @@ Based on:
 - [python v4.6.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v4.6.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.6.0-rc.2 - packages/orq-rc
+
+## 2026-03-12 06:27:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.752.0 (2.859.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.6.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.6.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.6.0-rc.3 - packages/orq-rc